Resolution of warnings at compile time

+++ lib/nrzl_encoder_impl.cc
@@ -50,10 +50,9 @@ namespace gr {
         throw std::out_of_range("nrzl encoder: the ratio samp rate on bit rate must to be integer and multiple of 2.");
       }
       positive = (float *)volk_malloc(d_interpolation * sizeof(float), volk_get_alignment());
-
       negative = (float *)volk_malloc(d_interpolation * sizeof(float), volk_get_alignment());
 
-      for (size_t i = 0; i < d_interpolation; i++)
+      for (int i = 0; i < d_interpolation; i++)
       {
         positive[i] = +1;
         negative[i] = -1;
@@ -75,7 +74,7 @@ namespace gr {
       const char *in = (const char *)input_items[0];
       float *out = (float *)output_items[0];
       int index_interpolation = 0;
-      for (size_t i = 0; i < noutput_items; i += d_interpolation)
+      for (int i = 0; i < noutput_items; i += d_interpolation)
       {
         if (in[index_interpolation] > 0)
         {
